+++ UI/ChallengeDetails/Tabs/Mechanics/Tabs/Damage/Tabs/DamageDoneByPartyMembers/Tabs/DamageDoneByPartyMember.lua
@@ -25,10 +25,10 @@ end
 @return[type=Frame]
 ]]
 function MyDungeonsBook:DamageDoneByPartyMemberFrame_Create(parentFrame, challengeId, unitId)
-    local ownCastsByPartyMemberFrame = self:TabContentWrapperWidget_Create(parentFrame);
+    local damageDoneByPartyMemberFrame = self:TabContentWrapperWidget_Create(parentFrame);
     local data = self:DamageDoneByPartyMemberFrame_GetDataForTable(challengeId, "ALL-DAMAGE-DONE-BY-PARTY-MEMBERS", unitId);
     local columns = self:DamageDoneByPartyMemberFrame_GetHeadersForTable(challengeId, unitId);
-    local table = self:TableWidget_Create(columns, 11, 40, nil, ownCastsByPartyMemberFrame, "all-damage-done-by-" .. unitId);
+    local table = self:TableWidget_Create(columns, 11, 40, nil, damageDoneByPartyMemberFrame, "all-damage-done-by-" .. unitId);
     table:SetData(data);
     table:RegisterEvents({
         OnClick = function(_, _, data, _, _, realrow, _, _, button)
@@ -54,7 +54,7 @@ function MyDungeonsBook:DamageDoneByPartyMemberFrame_Create(parentFrame, challen
             end
         end
     });
-    return ownCastsByPartyMemberFrame;
+    return damageDoneByPartyMemberFrame;
 end
 
 --[[--